Added DirectInput
Added DirectInput
Updated prompt "Start Quadcopter Flight"
Updated Blips on the mini map
Added option to hide HUD and radar
Added precision for floating variable in the menu
Updated long presses in the menu
Updated function GetTickCount to GetTickCount64
And other stuff

+++ LeFixDrone/TimeCycleManager.h
@@ -17,7 +17,7 @@ private:
 	static bool isFadingOut;
 	static bool isFadingIn;
 	static float strength;
-	static DWORD lastUpdate;
+	static ULONGLONG lastUpdate;
 	static float duration;
 };
